requirements.txt
aiohttp
anthropic
numpy
pandas
plotly
python-dotenv
streamlit
streamlit-autorefresh
web3
